WomCastTest class
